# Support implied end tags.