@Shinmera 2023-07-15
Thanks for your work.

Unfortunately I cannot merge this as-is, as it interferes in xml parsing modes that do not have the same HTML self-closing behaviour.