@aartaka 2023-07-15
Okay, what if the doctype is dynamically bound during parsing? Maybe make a toggle to force HTML/XML parsing?

What's the default parsing mode for Plump, actually? Is it XML?