@Shinmera 2023-07-15
There already is a toggle, it's binding the ``*tag-dispatchers*``.

The default parsing mode is a mix between the two, but that really doesn't matter for this.