@aartaka 2023-07-15
If I'm reading it right, then checking for XML tag presence in `*tag-dispatchers*` should be enough to say whether it's XML-ish parsing or no-XML HTML one?