@aartaka 2023-07-15
Something like:
``` lisp
(unless (any (lambda (dispatcher)
               (find (tag-dispatcher-name dispatcher) *tag-dispatchers* :key #'tag-dispatcher-name))
             *xml-tags*)
  #|...|#)
```