@Shinmera 2023-07-15
No, that would be an extremely ugly and leaky hack that would break custom dispatcher tables. The fundamental approach of this PR is not mergable.