@aartaka 2023-07-15
Okay, what about defining a fully custom default dispatcher (whatever that means, I don't yet understand all this infrastructure of dispatchers) for HTML tags, making it absolutely separate from XML?