@Shinmera 2023-09-17
I suppose you could install a custom default parser in the dispatch table for html.